requirements.txt
# Core agent stack
langchain>=0.2,<0.3
langchain-core>=0.2,<0.3
langgraph>=0.0.70,<0.2

# Model integrations
langchain-google-genai>=1.0.0,<2.0.0
google-generativeai>=0.5.0,<0.7.0

# Configuration & validation
python-dotenv>=1.0,<2.0
pydantic>=2.5,<3.0
